requirements.txt
python-telegram-bot>=20.0
langchain-core>=0.1.0
langchain-groq>=0.0.1
langchain-cohere>=0.0.1
requests>=2.28.0 